refactor: 디렉터리 정리 및 불필요한 코드 삭제
