+++ src/assets/default/people.json
