+++ src/components/Footer.jsx
@@ -2,7 +2,7 @@ import { useState } from "react";
 import { FaGithub } from 'react-icons/fa';
 import { FiLink } from 'react-icons/fi';
 import './style.css';
-import { relatedSites } from "../static/jsons";
+import { relatedSites } from "../assets/jsons";
 
 function Footer(props) {
     const [isActive, setIsActive] = useState(false);
